pyproject.toml
[project]
name = "talky"
version = "1.0.0"
description = "Voice-driven intelligent agent for Telegram with AI reasoning"
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Talky Team"}
]
dependencies = [
    "python-telegram-bot>=21.0.0",
    "openai>=1.12.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "asyncio-throttle>=1.0.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

